Merged changes for COP-20274

* commit '4f07e8850f1e9b6a00a2ee45d000a73c8b55583e':
  Incorporated review comments added decimal handling in cifssharehelper and nfsexporthelper removed matcher changes
  Handled decimal values for limits during quota directory creation
  fixed matchers
  added handling of decimal values for quota limits

+++ vipr-portal/com.iwave.isa.content/src/java/com/emc/sa/service/vipr/file/CreateCifsShareHelper.java
@@ -41,15 +41,15 @@ public class CreateCifsShareHelper {
 
     @Param(SHARE_COMMENT)
     protected String shareComment;
-    
-    @Param(value=SOFT_LIMIT,required=false)
-    protected Integer softLimit;
-    
-    @Param(value=ADVISORY_LIMIT,required=false)
-    protected Integer advisoryLimit;
-    
-    @Param(value=GRACE_PERIOD,required=false)
-    protected Integer gracePeriod;
+
+    @Param(value = SOFT_LIMIT, required = false)
+    protected Double softLimit;
+
+    @Param(value = ADVISORY_LIMIT, required = false)
+    protected Double advisoryLimit;
+
+    @Param(value = GRACE_PERIOD, required = false)
+    protected Double gracePeriod;
 
     @Bindable(itemType = FileStorageUtils.FileSystemACLs.class)
     protected FileStorageUtils.FileSystemACLs[] fileSystemShareACLs;
@@ -67,11 +67,12 @@ public class CreateCifsShareHelper {
     }
 
     public FileShareRestRep createCifsShare() {
-        int tempSoftLimit=(softLimit!=null)?softLimit:0;
-        int tempAdvisoryLimit=(advisoryLimit!=null)?advisoryLimit:0; 
-        int tempGracePeriod=(gracePeriod!=null)?gracePeriod:0;
-        
-        this.fileSystemId = FileStorageUtils.createFileSystem(project, virtualArray, virtualPool, shareName, sizeInGb, tempAdvisoryLimit,tempSoftLimit, tempGracePeriod);
+        int tempSoftLimit = (softLimit != null) ? softLimit.intValue() : 0;
+        int tempAdvisoryLimit = (advisoryLimit != null) ? advisoryLimit.intValue() : 0;
+        int tempGracePeriod = (gracePeriod != null) ? gracePeriod.intValue() : 0;
+
+        this.fileSystemId = FileStorageUtils.createFileSystem(project, virtualArray, virtualPool, shareName, sizeInGb, tempAdvisoryLimit,
+                tempSoftLimit, tempGracePeriod);
         FileStorageUtils.createCifsShare(this.fileSystemId, shareName, shareComment, null);
         return FileStorageUtils.getFileSystem(this.fileSystemId);
     }
